requirements.txt
allure_python_commons==2.13.1
jsonpath==0.82
pytest==6.2.4
python_jenkins==1.8.0
PyYAML
Requests==2.31.0
urllib3==1.26.15
